+++ Assets/Scripts/Sagas/InternalAuthenticate.cs
@@ -14,7 +14,7 @@ namespace CryptoQuest.Sagas
         public bool success;
         public string message;
         public int gold;
-        public int diamond;
+        public float diamond;
         public int soul;
         public long time;
         public CredentialResponse data;
